Rename packages
Group classes into packages base on their functionalities for easier management.

Let's move related classes into packages such as,
* gary.task
* gary.ui
* gary.gui
* gary.exception
