+++ src/main/java/gary/MainWindow.java
@@ -1,16 +1,13 @@
 package gary;
 
+import gary.gui.DialogBox;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
-import javafx.scene.control.Label;
 import javafx.scene.control.ScrollPane;
 import javafx.scene.control.TextField;
 import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
-import gary.Gary;
-import gary.DialogBox;
 /**
  * Controller for MainWindow. Provides the layout for the other controls.
  */
