+++ src/main/java/gary/task/Task.java
@@ -1,4 +1,4 @@
-package gary;
+package gary.task;
 
 /**
  * Represents a task
