+++ src/main/java/gary/task/ToDo.java
@@ -1,5 +1,4 @@
-package gary;
-import gary.Task;
+package gary.task;
 
 /**
  * Represents a Task to be done
