+++ src/test/java/gary/ParserTest.java
@@ -1,10 +1,12 @@
 package gary;
-import gary.TaskList;
-import gary.Parser;
+import gary.exception.GaryException;
+import gary.ui.Parser;
+import gary.ui.TaskList;
 import org.junit.jupiter.api.Test;
+
 import java.util.ArrayList;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.fail;
 
 /**
  * Test for correctness of gary.Parser class
@@ -21,7 +23,7 @@ public class ParserTest {
             Parser p = new Parser();
             TaskList t = new TaskList(new ArrayList<String>());
             assertEquals("Bye, have a productive day!", p.parse("bye", t));
-        } catch (gary.GaryException e) {
+        } catch (GaryException e) {
             System.out.println("testBye: error");
         }
     }
