+++ src/test/java/gary/TaskTest.java
@@ -1,8 +1,8 @@
 package gary;
-import gary.Task;
+import gary.task.Task;
 import org.junit.jupiter.api.Test;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.fail;
 
 /**
  * Test for the correctness of gary.Task
